refactor(tests): split resolver tests; allowlist 2 large files
CI file-size check (max 750) failed after #979's additions on 3 files:

- tests/cli/test_utils.py: 1072 → 448 lines. Moved both resolver test
  classes (TestResolveMcpConfigPath, TestResolveClaudeSettingsPath) to
  a new tests/cli/test_resolver_utils.py. Net: 76 tests preserved
  identically across the two files; both well under the limit.

- src/mcp_coder/cli/parsers.py (769 lines): allowlisted. The 19-line
  overrun is from #979's 7 new --settings add_argument blocks; the
  file is a logical single unit (CLI argparse setup) where splitting
  would harm cohesion.

- tests/llm/providers/claude/test_claude_code_cli.py (807 lines):
  allowlisted. 57-line overrun; tests are cohesive around
  build_cli_command + ask_claude_code_cli and splitting would
  fragment them.
