+++ .large-files-allowlist
@@ -9,6 +9,7 @@ docs/configuration/config.md
 docs/processes-prompts/development-process.md
 src/mcp_coder/cli/commands/coordinator/commands.py
 src/mcp_coder/cli/commands/verify.py
+src/mcp_coder/cli/parsers.py
 tools/safe_delete_folder.py
 src/mcp_coder/llm/mlflow_logger.py
 src/mcp_coder/prompt_manager.py
@@ -21,6 +22,7 @@ tests/icoder/test_app_pilot.py
 tests/cli/commands/test_prompt.py
 tests/cli/commands/test_verify_orchestration.py
 tests/cli/test_main.py
+tests/llm/providers/claude/test_claude_code_cli.py
 tests/llm/test_interface.py
 tests/utils/test_folder_deletion.py
 tests/utils/github_operations/issues/test_get_branch_with_pr_fallback.py
